Add client for communication
The ChartMogul API uses HTTP Basic Authentication. This means that
with every request we must provide Account Token and Secret Key as
basic auth credentials. It also expects request body to be in json

This commit adds an API Client that will set those authentication
headers, and it will also prepare the request body as expected. It
will also parse the API response to OpenStruct object.