Add configurable response type
We have been using `OpenStruct` to parse the API responses to
an object. Normally `OpenStruct` is slower in performance. The
amount of data our API client will be dealing is not that
large so technically it should not effect at all.

But let's change it be configurable so user can use it based on
their necessity.